compatible with empty gene file

+++ lib/Annotation/WebGestaltR.pm
@@ -55,8 +55,16 @@ sub perform {
     print $pbs " 
 if [[ ! -s $final_file || ! -d $final_file ]]; then
   cd $cur_dir 
-  R --vanilla -f $script --args $organism $sample_name $inputFile . $interestGeneType $referenceSet
-  rm */*.tar.gz
+  if [[ -f $inputFile ]]; then
+    if [[ -s $inputFile ]]; then
+      R --vanilla -f $script --args $organism $sample_name $inputFile . $interestGeneType $referenceSet
+      rm */*.tar.gz
+    else
+      echo \"Empty gene file\" > ${sample_name}.empty
+    fi 
+  else
+    echo \"No gene file exist.\"
+  fi
 fi
 
 ";
@@ -73,7 +81,6 @@ sub result {
   my $result       = {};
   for my $sample_name ( sort keys %$comparisons ) {
     my $cur_dir = scalar( keys %$comparisons ) == 1 ? $result_dir : create_directory_or_die( $result_dir . "/$sample_name" );
-    my $finalFile = "Project_" . $sample_name . "_geneontology_Biological_Process/Report_" . $sample_name . "_geneontology_Biological_Process.html";
     my @result_files = ();
     push( @result_files, "$cur_dir/Project_${sample_name}_geneontology_Biological_Process/enrichment_results_${sample_name}_geneontology_Biological_Process.txt" );
     push( @result_files, "$cur_dir/Project_${sample_name}_geneontology_Cellular_Component/enrichment_results_${sample_name}_geneontology_Cellular_Component.txt" );
